fix script and test

+++ script/DeploySetupConfig.s.sol
@@ -17,4 +17,4 @@ uint256 constant DEBT_GAS_COMPENSATION = 2e18; // 2 SAT
 
 uint32 constant SP_CLAIM_START_TIME = 4_294_967_295; // max uint32
 uint256 constant SP_ALLOCATION = 0; // 0 initial allocation
-uint256 constant SP_REWARD_RATE = 0; // 0 initial reward rate
+uint128 constant SP_REWARD_RATE = 0; // 0 initial reward rate
